InterruptedException to Exception change
We're observing the enormous number of crashes (coming mostly from Firebase's stack trace) in PluginMarker.java file. The line 167 is generating an Exception that is not being predicted by the Try-Catch where is inserted (lines 149 - 185), which is specifically only catching InterruptedException types. Simply replacing "InterruptedException" with "Exception" will solve the problem, as then the catch will be able to catch all Exception types and avoid the crash.